% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalua_secuencia.R
\name{evalua_secuencia}
\alias{evalua_secuencia}
\title{Evalúa una secuencia de nucleótidos}
\usage{
evalua_secuencia(seq)
}
\arguments{
\item{seq}{Un vector tipo caracter}
}
\value{
Un vector tipo caracter resultado de la evaluación
}
\description{
Esta función toma como entrada una secuencia
    de nucléotidos y evalúa si es tipo DNA o RNA.
}
\examples{
seq <- "attataggcttt"
evalua_secuencia(seq)
}
