% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_menos_min.R
\name{max_menos_min}
\alias{max_menos_min}
\title{Calcula el rango}
\usage{
max_menos_min(x)
}
\arguments{
\item{x}{Un vector numérico}
}
\value{
Un valor numérico
}
\description{
Esta función calcula el rango de una serie de valores
    numéricos de un vector.
}
\examples{
x_num <- seq(1:100)
max_menos_min(x_num)
}
