% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_heatmap.R
\name{subset_heatmap}
\alias{subset_heatmap}
\title{Title}
\usage{
subset_heatmap(x, mediciones = NULL, grupos = NULL)
}
\arguments{
\item{x}{matriz de genes x mediciones}

\item{mediciones}{character() mediciones a seleccionar}

\item{grupos}{character() genes a seleccionar}
}
\value{
objeto "ComplexHeatmap::Heatmap"
}
\description{
Title
}
\examples{
#crear matriz
mi_matriz <- matrix(rnorm(100), nrow = 10)
rownames(mi_matriz) <- paste0("medicion_",letters[1:10])
colnames(mi_matriz) <- paste0("grupo_",letters[1:10])

#librerias
library(ComplexHeatmap)
subset_heatmap(
  mi_matriz,
  mediciones = c("medicion_a", "medicion_b", "medicion_c"),
  grupos = c("grupo_d","grupo_e","grupo_f"))

}
